chore: move dto and entities to corresponding directories
