+++ src/addresses/entities/address.entity.ts
@@ -1,4 +1,4 @@
-import { User } from 'src/dto/users/user.entity';
+import { User } from 'src/users/entities/user.entity';
 import {
   Column,
   Entity,
