+++ src/app.module.ts
@@ -4,11 +4,11 @@ import { Address } from './addresses/entities/address.entity';
 import { AddressesModule } from './addresses/addresses.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { User } from './dto/users/user.entity';
+import { User } from './users/entities/user.entity';
 import { WorkingHours } from './working-hours/entites/working-hours.entity';
 import { UsersModule } from './users/users.module';
 import { PhonesModule } from './phones/phones.module';
-import { Phone } from './dto/phones/phone.entity';
+import { Phone } from './phones/entities/phone.entity';
 import { Organizer } from './organizers/entities/organizer.entity';
 import { OrganizersModule } from './organizers/organizers.module';
 import { Visit } from './visits/entities/visit.entity';
