+++ src/phones/dto/create-phone.dto.ts
@@ -1,6 +1,6 @@
 import { OmitType } from '@nestjs/swagger';
 import { IsUUID } from 'class-validator';
-import { Phone } from './phone.entity';
+import { Phone } from '../entities/phone.entity';
 
 export class CreatePhoneDto extends OmitType(Phone, ['user'] as const) {
   @IsUUID()
