+++ src/phones/phones.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Phone } from 'src/dto/phones/phone.entity';
+import { Phone } from 'src/phones/entities/phone.entity';
 import { UsersModule } from 'src/users/users.module';
 import { PhonesController } from './phones.controller';
 import { PhonesService } from './phones.service';
