+++ src/users/dto/create-user.dto.ts
@@ -1,5 +1,5 @@
 import { OmitType } from '@nestjs/swagger';
-import { User } from './user.entity';
+import { User } from '../entities/user.entity';
 
 export class CreateUserDto extends OmitType(User, [
   'id',
