+++ src/users/users.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from 'src/dto/users/user.entity';
+import { User } from 'src/users/entities/user.entity';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
